+++ interpreter/alloc.h
@@ -1,6 +1,6 @@
 /*
  * nix operating system project tre interpreter
- * Copyright (c) 2005-2007 Sven Klose <pixel@copei.de>
+ * Copyright (c) 2005-2008 Sven Klose <pixel@copei.de>
  *
  * Memory allocation.
  */
@@ -8,7 +8,9 @@
 #ifndef TRE_ALLOC_H
 #define TRE_ALLOC_H
 
-#include <wordexp.h>
+#include "ctype.h"
+
+#include <sys/types.h>
 
 extern treptr trealloc_used;
 extern treptr trealloc_unused;
@@ -18,6 +20,6 @@ extern void trealloc_free (void *);
 
 extern void * trealloc_item (void * handle, void * map, void * end);
 extern void trealloc_free_item (void * handle, void * item, void * map, void * end);
-extern void * trealloc_item_init (void * map, unsigned num_items, unsigned item_size);
+extern void * trealloc_item_init (void * map, ulong num_items, ulong item_size);
 
 #endif
