+++ interpreter/argument.h
@@ -11,8 +11,8 @@
 extern void trearg_get2 (treptr *car, treptr *cdr, treptr args);
 extern treptr trearg_get (treptr args);
 
-extern treptr trearg_correct (unsigned argnum, int type, treptr, const char * descr);
-extern treptr trearg_typed (unsigned argnum, int type, treptr, const char * descr);
+extern treptr trearg_correct (ulong argnum, int type, treptr, const char * descr);
+extern treptr trearg_typed (ulong argnum, int type, treptr, const char * descr);
 
 extern void trearg_expand (treptr *rvars, treptr *rvals,
                             treptr argdef, treptr args,
